ticket3502
git-svn-id: http://repos.luxurylink.com/svn/repos/toolbox/trunk@4636 bd3a7434-8174-4811-bcf6-7b9fed35a952

+++ app/views/users/email.ctp
@@ -1,47 +1,54 @@
 <?
 
 echo "<b>$email</b>";
-echo "<br><br>green=keep, yellow=make inactive, red=delete - ";
-echo "<a href='?email=$email&process=1' onclick='return confirm(\"Are you sure?\");'><b>Do it</b></a>";
+echo "<br><br>";
+?>
+<b>How it works.</b>
+<br><br>
+Emails that have userId's that all have matching rows in userSiteExtended OR no matching rows in userSiteExtended are processed first.  This makes userSiteExtended no longer a criteria for the rest of the script.
+<br><br>Then set the primary userId by:<br><br>
+1. user row with most recent modifyDateTime (login) that has a ticketId<br>
+2. or the most recent (non null) modifyDateTime<br>
+3. or the most recent userId<br><br>
+<?
+echo "<a href='?email=$email&process=1' onclick='return confirm(\"Are you sure?\");'><b>Process</b></a>";
 echo "<br><br>";
 
-echo "<table border=1 cellpadding=4 cellspacing=0 style='width:750px;'>";
-echo "<tr>";
-echo "<td>userId</td>";
-echo "<td>inactive</td>";
-echo "<td>createDateTime</td>";
-echo "<td>modifyDateTime</td>";
-echo "<td>ticketId</td>";
-echo "<td>userSiteExtendedId</td>";
-echo "</tr>";
+if (count($rowArr)>0){
 
-krsort($rowArr);
-foreach($rowArr as $row){
-
+	echo "<table border=1 cellpadding=4 cellspacing=0 style='width:750px;'>";
 	echo "<tr>";
-	echo "<td>";
-	echo "<a href=/users/edit/".$row['userId'].">";
-	echo "<span style='background-color:";
-	if (in_array($row['userId'], $keepArr[$row['email']])){
-		echo "green;color:#ffffff;";
-	}else if (in_array($row['userId'],$delIdArr)){
-		echo "#e80000;color:#ffffff;";
-	}else{
-		echo "white";
-	}
-	echo ";'>";
-	echo $row['userId']."</span></a>";
-	echo "</td>";
-	echo "<td>".$row['inactive']."</td>";
-	echo "<td>".$row['createDateTime']."</td>";
-	echo "<td>".$row['modifyDateTime']."</td>";
-	echo "<td>";
-	echo "<span style='background-color:";
-	echo in_array($row['userId'], $inactiveArr[$row['email']])?"yellow":"white";
-	echo ";'>";
-	echo $row['ticketId']."</span></td>";
-	echo "<td>".$row['userSiteExtendedId']."</td>";
+	echo "<td>userId</td>";
+	echo "<td>inactive</td>";
+	echo "<td>createDateTime</td>";
+	echo "<td>modifyDateTime</td>";
+	echo "<td>ticketId</td>";
+	echo "<td>userSiteExtendedId</td>";
 	echo "</tr>";
 
-}	
-echo "</table>";
+	foreach($rowArr as $userId=>$row){
+
+		echo "<tr>";
+		echo "<td>";
+		echo "<a href=/users/edit/".$row['userId'].">";
+		echo $row['userId']."</span></a>";
+		echo "</td>";
+		echo "<td>".$row['inactive']."</td>";
+		echo "<td>".$row['createDateTime']."</td>";
+		echo "<td>".$row['modifyDateTime']."</td>";
+
+		echo "<td>";
+		echo isset($row['ticketId'])?$row['ticketId']:' - ';
+		echo "</td>";
+
+		echo "<td>";
+		echo isset($row['userSiteExtendedId'])?$row['userSiteExtendedId']:' - ';
+		echo "</td>";
+		echo "</tr>";
+
+	}	
+	echo "</table>";
+
+}else{
+	echo "No rows found for $email";
+}
